% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_react_funcs.r
\name{add.CIs}
\alias{add.CIs}
\title{Define a function to take a table and add in binomial estimates and
CIs}
\usage{
add.CIs(tab, method = "wilson")
}
\description{
Define a function to take a table and add in binomial estimates and
CIs
}
