% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attack_rate.r
\name{attack_rate}
\alias{attack_rate}
\title{Solves the classic relationship between R0 and cumulative attack rate}
\usage{
attack_rate(R0 = 1.8)
}
\arguments{
\item{R0}{basic reproductive number}
}
\description{
Solves the classic relationship between R0 and cumulative attack rate
}
\examples{
attack_rate(1.8)
attack_rate(4.0)
attack_rate(3.0)
}
