% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_practical.R
\name{calc_forecast_accuracy}
\alias{calc_forecast_accuracy}
\title{calculate proportion of forecast points which are within threshold}
\usage{
calc_forecast_accuracy(forecast_df, tolerance = 0.25)
}
\arguments{
\item{forecast_df}{forecast data frame returned by 
\code{extract_forecasted_points} or \code{extract_forecasted_points_seir}}

\item{tolerance}{numeric vector of length 1: parameter specifying range
within which the forecasted incidence should fall.  For example, if 
tolerance = 0.25, a forecasted point is deemed to be accurate if the 
forecasted incidence is between 75% and 125% of the observed incidence.}
}
\value{
numeric vector of length 1: the proportion of forecast points 
which are within a given percentage threshold of the observed incidence
}
\description{
\code{calc_forecast_accuracy} calculates the proportion of forecast points 
which are within a given percentage threshold of the observed incidence
}
