% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_seir.R
\name{comp_seir}
\alias{comp_seir}
\title{Solves the SEIR model}
\usage{
comp_seir(De = 7, Tg = 15, R0 = 1.8, N = 1e+06, I0 = 10,
  trickle = 0, trickleStart = 0, reprate = 1, dt = 1, R1 = 1,
  t1 = 999, R2 = 1, t2 = 9999, noTimeSteps = 360, noReals = 1,
  A = 0, deterministic = TRUE)
}
\arguments{
\item{De}{duration latent}

\item{Tg}{generation time}

\item{R0}{Basic reproductive number}

\item{N}{size of the population}

\item{I0}{number of initial infectious individuals}

\item{trickle}{expected number of trickle infections per day}

\item{trickleStart}{start time for trickle infection}

\item{reprate}{reporting rate}

\item{dt}{time step}

\item{R1}{second value for the reproductive nuber}

\item{t1}{time of change to the second R value}

\item{R2}{third value for the reproductive number}

\item{t2}{time of change to the third reproductive number}

\item{noTimeSteps}{number of timesteps}

\item{noReals}{number of realizations, NOT YET IMPLEMENTED}

\item{A}{coefficient of seasonaliity}

\item{deterministic}{mode of solutions}
}
\value{
A list of two elements. The first is the incidence of infection
and the second is the timepoints to which the incidence refers.
}
\description{
Creates both deterministic and stochastic solutions to a common
epidemic model.
}
\examples{
comp_seir()
comp_seir(R0=0.999)
}
