% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_svir.r
\name{comp_svir}
\alias{comp_svir}
\title{Solves the SVIR model}
\usage{
comp_svir(
  Tg = 2.8,
  R0 = 1.8,
  N = 1e+05,
  I0 = 1,
  V0 = 0,
  dt = 0.1,
  startVac = 9999,
  vacRate = 0,
  noTimeSteps = 10 * 180,
  noReals = 1,
  deterministic = TRUE,
  doPlot = FALSE
)
}
\arguments{
\item{Tg}{generation time}

\item{R0}{Basic reproductive number}

\item{N}{size of the population}

\item{I0}{number of initial infectious individuals}

\item{dt}{time step}

\item{noReals}{number of realizations, NOT YET IMPLEMENTED}

\item{deterministic}{mode of solutions}

\item{A}{coefficient of seasonaliity}
}
\value{
A list of two elements. The first is the incidence of infection
and the second is the timepoints to which the incidence refers.
}
\description{
Creates both deterministic and stochastic solutions to a common
epidemic model.
}
\details{
Details go here XXX new edit YYYY
}
\examples{
comp_svir()
comp_svir(R0=0.999)
}
