% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_excess_econ}
\alias{country_excess_econ}
\title{Economist weekly COVID-19 excess mortality data for 88 weeks from week 1 2020}
\format{
A data 88 rows and 238 columns
\describe{
  \item{country}{Country names are row labels}
  \item{week}{Weeks starting from week 1 2020 are row labels}
}
}
\source{
<https://raw.githubusercontent.com/owid/covid-19-data/master/public/data/excess_mortality/excess_mortality_economist_estimates.csv>
}
\usage{
country_excess_econ
}
\description{
Downloaded from our world in data
}
\keyword{datasets}
