% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_branch.r
\name{cov_branch}
\alias{cov_branch}
\title{Simulates a stochastic epidemic being exported by air travel}
\usage{
cov_branch(
  nreal = 2,
  ninf = 100,
  tinfmax = 999,
  seed = 10,
  popN = -1,
  gtd = dpois(0:20, 8.1),
  off = dpois(0:10, 2.6),
  ipd = dpois(0:10, 5.8),
  pass_per_day = 3301,
  R0 = -1,
  k = 1,
  popsize = 1.9e+07,
  casetypes = c(1, 1, 1)
)
}
\description{
Function to simulate an ncov-like brtanching process.
This might be useful for the border crossing work
Need to go through this and see if I cam make it more 
relevant to omicron
}
\examples{
cov_branch()
}
