% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_hybrid.r
\name{cov_hybrid}
\alias{cov_hybrid}
\title{Solves a compartmental model for SARS-CoV-2 like model}
\usage{
cov_hybrid(
  R0 = 2,
  Rp = 1.5,
  Rover = 0.9,
  vecN = 6.6e+07 * c(0.25, 0.25, 0.25, 0.25),
  vecI0 = rep(5000/length(vecN), length(vecN)),
  vecInfNess = rep(1, length(vecN)),
  vecSusTy = rep(1, length(vecN)),
  vecPS = rep(0.1, length(vecN)),
  vecPM = rep(0.1, length(vecN)),
  vecPA = rep(0.1, length(vecN)),
  matCt = matrix(1/length(vecN), nrow = length(vecN), ncol = length(vecN)),
  matCtClosure = matrix(1/length(vecN), nrow = length(vecN), ncol = length(vecN)),
  scLim = c(99999, 99999),
  vecTcalc = seq(0, 720, 0.1),
  vecRtrel = rep(1, length(vecTcalc)),
  D_E = 2,
  D_I1 = 3,
  D_I2 = 3,
  D_HR = 20,
  D_HD = 23,
  D_ICU = rep(7, length(vecN)),
  deterministic = TRUE,
  nReals = 2,
  trig_pres = 99999999,
  icu_cap = 0.9999,
  plot = FALSE,
  trickle = 0,
  sevBchange = FALSE
)
}
\arguments{
\item{Description}{here}
}
\value{
Add the return information here

XXXX This function not fully implemented yet XXXX
}
\description{
Basic compartmental model for age-structured covid. Can run as
deterministic or stochastic on fixed time step. Has severity structure
with finite capacity for severe cases. Useful edge-case tests implemented
at the end of the function script in an "if (FALSE) {...}" block. These
could be added to a model intro vingette at some point.
}
