% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_sc.r
\name{cov_sc}
\alias{cov_sc}
\title{Explores school closure options for a viral respiratory pandemic}
\usage{
cov_sc(
  vecR0 = seq(1.5, 3, 0.5),
  vecTStart = seq(21, 21 + 2 * 7, 7),
  vecDuration = seq(14, 14 + 3 * 28, 28),
  vecSusChild = c(1, 0.333)
)
}
\description{
Used in Feb 2020.
}
