% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_sitrep_tab.r
\name{cov_sitrep_tab}
\alias{cov_sitrep_tab}
\title{export
Define a function to make tables for different fields in the sitrep
data}
\usage{
cov_sitrep_tab(df, fldname = "new_dx_total", long = TRUE, propNAs = TRUE)
}
\description{
export
Define a function to make tables for different fields in the sitrep
data
}
