% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_practical.R
\name{extract_forecasted_points_seir}
\alias{extract_forecasted_points_seir}
\title{extract forecasted points for SEIR model}
\usage{
extract_forecasted_points_seir(R_0, incidence_data, current_week, starting_week,
  weeks_ahead)
}
\arguments{
\item{R_0}{numeric vector of length 1: basic reproduction number}

\item{incidence_data}{data frame extracted by \code{extract_incidence}}

\item{current_week}{numeric vector of length 1: week number of the current week}

\item{starting_week}{numeric vector of length 1: 
guess for week number when the epidemic started. Use data from starting week
to current week to forecast}

\item{weeks_ahead}{numeric vector of length 1: number of weeks to forecast ahead}
}
\value{
data frame with predicted points and original data
}
\description{
\code{extract_forecasted_points_seir} takes the R_0
returned by \code{fit_seir} and extracts model predictions
}
