% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_practical.R
\name{extract_incidence}
\alias{extract_incidence}
\title{extracts the incidence data for a given country and year}
\usage{
extract_incidence(flu_data, country_code = "ISR", year = 2016)
}
\arguments{
\item{flu_data}{matrix with flu incidence data}

\item{country_code}{character vector of length 1: 3-letter country code}

\item{year}{numeric vector of length 1: year to extract}
}
\value{
data frame with three columns:
t: numeric vector of length 52/53 (depending on the year): 
index running from 1 to 52/53
time_name: character vector of the same length as t:
week numbers of the data we've extracted
incidence: numeric vector of the same length as t:
incidence for those weeks
}
\description{
\code{extract_incidence} takes the flu incidence matrix and extracts
the incidence for a given flu season and a given country.
}
