% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_practical.R
\name{extract_predicted_points}
\alias{extract_predicted_points}
\title{extract predicted points from linear regression object}
\usage{
extract_predicted_points(lm_output, incidence_data, weeks_ahead,
  log_transform)
}
\arguments{
\item{lm_output}{linear regression object returned by \code{linear_regression}}

\item{incidence_data}{incidence data extracted by \code{extract_incidence}}

\item{weeks_ahead}{numeric vector of length 1: number of weeks to predict ahead}

\item{log_transform}{logical vector of length 1: if TRUE, linear 
regression was performed on log transformed data, otherwise linear regression
was performed on original data}
}
\value{
data frame with predicted points and original data
}
\description{
\code{extract_predicted_points} takes the linear regression object
returned by \code{linear_regression} and extracts model predictions
}
