% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_practical.R
\name{extract_subset_prediction}
\alias{extract_subset_prediction}
\title{extract subset of prediction data frame containing predictions
#' 
\code{calc_prediction_accuracy} calculates the proportion of prediction points 
which are within a given percentage threshold of the observed incidence}
\usage{
extract_subset_prediction(prediction_df)
}
\arguments{
\item{prediction_df}{prediction data frame returned by 
\code{extract_predicted_points} or \code{extract_predicted_points_seir}}
}
\value{
data frame: incidence and prediction for time points where data is 
available
}
\description{
extract subset of prediction data frame containing predictions
#' 
\code{calc_prediction_accuracy} calculates the proportion of prediction points 
which are within a given percentage threshold of the observed incidence
}
