% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_practical.R
\name{extract_week_index}
\alias{extract_week_index}
\title{extract index of week number from week names}
\usage{
extract_week_index(week_no, time_name)
}
\arguments{
\item{week_no}{numeric vector of length 1: week number of the week of interest}

\item{time_name}{character vector with names of weeks}
}
\value{
numeric vector of length 1: index
}
\description{
\code{extract_week_index} extract index of week number from week names
}
