% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allData.r
\docType{data}
\name{fluIliCountryData}
\alias{fluIliCountryData}
\title{Simulation model example data frame}
\description{
This object contains synthetic data on the location of farms in Kent, UK.
These are not real locations but they have a similar density distribution.
These data are intended to be used in a teaching example about
simulation models of infectious disease.
}
\references{
These incidence data are from publically available WHO syndromic data
\url{https://www.who.org}
}
\author{
Steven Riley \email{sr@stevenriley.net}
}
