% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{homevisits}
\alias{homevisits}
\title{A panel dataset of residential fires in 153 rescue service confederations in Sweden (2000-2015).}
\format{A data frame with N=153 units observed over T=17 years, containing the following variables:
\describe{
  \item{year}{actual year}
  \item{id}{unit id}
  \item{rtjf_15}{unit name}
  \item{pop}{population size}
  \item{fires}{number of residential fire events}
  \item{treat}{treatment unit dummy variable}
  \item{post}{post-intervention dummy, coded as 1 if year>=2010}
}}
\usage{
homevisits
}
\description{
A panel dataset of residential fires in 153 rescue service confederations in Sweden (2000-2015).
}
\examples{
\dontrun{
#This code replicates the case study results from the paper (Bonander, in preparation).

data(homevisits) #Open the data

#Find the optimal value for k and estimate effects

firmod <- idd(eventvar="fires",
             popvar="pop",
             idvar="id",
             timevar="year",
             postvar="post",
             treatvar="treat",
             names="rtjf_15",
             data=homevisits)

firmod$id_controls #Show the selected controls

idd.kplot(firmod) #Plot the k-minimization function (figure 2)

#Reproduce figure 3 using grid and gridExtra

A <- idd.cfplot(firmod, mult=100000)
B <- idd.gplot(firmod, mult=100000)
B <- idd.pplot(firmod)

require(grid)
require(gridExtra)

grid.arrange(A, B, C, ncol=1)

#Perform the placebo studies (fair warning: 1000 iterations will take a while; expect 15-30 min)

set.seed(12049135) ##Set seed to replicate the results in the paper

placebt <- iddplacebo(eventvar="fires",
                     popvar="pop",
                     idvar="id",
                     timevar="year",
                     postvar="post",
                     treatvar="treat",
                     data=homevisits,
                     iter=1000)

#Reproduce figure 4

A2 <- iddplacebo.hist(placebt, convert=F, quantile=T)
B2 <- iddplacebo.hist(placebt, convert=T, quantile=T)
C2 <- iddplacebo.ecdf(placebt, convert=T, quantile=T)
grid.arrange(A2, B2, C2, ncol=1)

#Obtain placebo-based confidence intervals and p-values

pci <- placeboci(placebt, alpha=0.05)

#Reproduce figure 5

A1 <- placebo.ciplot(pci)
B1 <- placebo.pplot(pci)
grid.arrange(A1, B1, ncol=1)

##Other results and features

#Find the probability of finding a parametric p-value <=0.05 in the untreated units

mean(as.numeric(placebt$Resdata$param_p<=0.05))

#Generate a Synthetic control-style "spaghetti plot" to view the placebo effects

iddplacebo.spaghetti(placebt)

#Find the RMSE ratio for the treated unit

placebt$Treat.ratio

#Look at the raw data (without standardizing the matched controls to the level of the treated unit)

idd.dplot(firmod)

#Add a line for the entire donor pool for comparison

idd.dplot(firmod, donor=T)
}
}
\keyword{datasets}
