\name{idd-package}
\alias{idd-package}
\alias{idd}
\docType{package}
\title{
Infectious disease dynamics
}
\description{
A set of functions to help better understand the transmission of infectious diseases. 
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{idd}
\packageIndices{idd}
This package provides a variety of functions to help investigate infectious disease dynamics. The functions fall into }
\author{
Steven Riley
}
\references{
~~ Literature or other references for background information ~~
}
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%% ~~ the R documentation directory ~~
\keyword{ package }
%% \seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%% }
%% \examples{
%% ~~ simple examples of the most important functions ~~
%% }
