% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idd.cfplot.R
\name{idd.cfplot}
\alias{idd.cfplot}
\title{Plot the counterfactual.}
\usage{
idd.cfplot(x, mult = 1e+05, donor = FALSE)
}
\arguments{
\item{x}{Name of the \code{idd} output object.}

\item{mult}{Multiplier for the rates (default=100000).}

\item{donor}{Plot a counterfactual based on the entire donor pool as well? Defaults to \code{FALSE}.}
}
\value{
Returns a ggplot of the estimated counterfactual based on the selected control group.
}
\description{
\code{idd.cfplot} takes an output object after \code{idd} and plots the counterfactual using \code{ggplot2}.
}
\examples{
\dontrun{
data(simpanel)
idd.out <- idd(eventvar="y",
             popvar="pop",
             idvar="age",
             timevar="time",
             postvar="post",
             treatvar="treat",
             data=simpanel)
plot.out <- idd.cfplot(idd.out)
}
}
