% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idd.dplot.R
\name{idd.dplot}
\alias{idd.dplot}
\title{Plot the raw data.}
\usage{
idd.dplot(x, mult = 1e+05, donor = FALSE)
}
\arguments{
\item{x}{Name of the \code{idd} output object.}

\item{mult}{Multiplier for the rates (default=100000).}

\item{donor}{Plot data from the entire donor pool as well? Defaults to \code{FALSE}.}
}
\value{
Returns a ggplot of raw data (without standardization and removal of fixed effects).
}
\description{
\code{idd.dplot} takes an output object after \code{idd} and plots the raw data using \code{ggplot2}.
}
\examples{
\dontrun{
data(simpanel)
idd.out <- idd(eventvar="y",
             popvar="pop",
             idvar="age",
             timevar="time",
             postvar="post",
             treatvar="treat",
             data=simpanel)
plot.out <- idd.dplot(idd.out)
}
}
