% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idd.gplot.R
\name{idd.gplot}
\alias{idd.gplot}
\title{Plot the estimated effect.}
\usage{
idd.gplot(x, mult = 1e+05)
}
\arguments{
\item{x}{Name of the \code{idd} output object.}

\item{mult}{Multiplier for the rates (default=100000).}
}
\value{
Returns a ggplot of the estimated time-varying effects based on the selected control group.
}
\description{
\code{idd.gplot} takes an output object after \code{idd} and plots the estimated effect using \code{ggplot2}.
}
\examples{
\dontrun{
data(simpanel)
idd.out <- idd(eventvar="y",
             popvar="pop",
             idvar="age",
             timevar="time",
             postvar="post",
             treatvar="treat",
             data=simpanel)
plot.out <- idd.gplot(idd.out)
}
}
