% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idd.kplot.R
\name{idd.kplot}
\alias{idd.kplot}
\title{Plot the k-minimization function.}
\usage{
idd.kplot(x)
}
\arguments{
\item{x}{Name of the \code{idd} output object.}
}
\value{
Returns a ggplot of the k-minimization function used to select the optimal value of k.
}
\description{
\code{idd.cfplot} takes an output object after \code{idd} and plots the k-minimization function using \code{ggplot2}.
}
\examples{
\dontrun{
data(simpanel)
idd.out <- idd(eventvar="y",
             popvar="pop",
             idvar="age",
             timevar="time",
             postvar="post",
             treatvar="treat",
             data=simpanel)
plot.out <- idd.kplot(idd.out)
}
}
