% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idd.pplot.R
\name{idd.pplot}
\alias{idd.pplot}
\title{Plot the p-values.}
\usage{
idd.pplot(x, alpha = 0.05)
}
\arguments{
\item{x}{Name of the \code{idd} output object.}

\item{alpha}{Set alpha value for the significance region (defaults to 0.05).}
}
\value{
Returns a ggplot of the estimated time-varying p-values for the effect based on the selected control group.
}
\description{
\code{idd.gplot} takes an output object after \code{idd} and plots time-varying p-values for the effect using \code{ggplot2}.
}
\examples{
\dontrun{
data(simpanel)
idd.out <- idd(eventvar="y",
             popvar="pop",
             idvar="age",
             timevar="time",
             postvar="post",
             treatvar="treat",
             data=simpanel)
plot.out <- idd.pplot(idd.out)
}
}
