% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idd.R
\name{idd}
\alias{idd}
\title{Identify controls and estimate intervention effects.}
\usage{
idd(eventvar, popvar, treatvar, postvar, timevar, idvar, names = NULL,
  print = TRUE, data, mult = 1e+05)
}
\arguments{
\item{eventvar}{Name of the event count variable.}

\item{popvar}{The person-time (or another exposure) variable.}

\item{treatvar}{The treatment group indicator dummy (0 for controls, 1 for treated units).}

\item{postvar}{The post-period indicator dummy (0 for all time points in the pre-intervention period, 1 for all time points in the post-period)}

\item{timevar}{The time variable (can be coded on any scale).}

\item{idvar}{The panel id variable.}

\item{names}{A variable containing unit names (optional). Changes the output of id.selected to show unit names of selected controls.}

\item{print}{Print results? Default=TRUE.}

\item{data}{A long-form panel dataset containing the supplied variables.}

\item{mult}{Multiplier for the rates produced by print (default=100000).}
}
\value{
Returns a list containing the following elements:
\code{$Resdat}: a data frame containing the results.
\code{$cv_errors}: a data frame containing the cross-validation errors for the k-minimizing function.
\code{$supp_stats}: a data frame containing supplementary statistics (average effects, relative effects, p-values etc).
\code{$id_controls}: a data frame containing information on the selected controls.
}
\description{
\code{idd} takes a long-form panel dataset containing event count and exposure data, obtains the optimal controls and estimates an incidence difference-in-differences model.
}
\examples{
\dontrun{
data(simpanel)
idd.out <- idd(eventvar="y",
             popvar="pop",
             idvar="age",
             timevar="time",
             postvar="post",
             treatvar="treat",
             data=simpanel)
}
}
