% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iddplacebo.R
\name{iddplacebo}
\alias{iddplacebo}
\title{Run placebo studies on untreated units and estimate placebo p-values.}
\usage{
iddplacebo(eventvar, popvar, treatvar, postvar, timevar, idvar, data,
  iter = 10)
}
\arguments{
\item{eventvar}{Name of the event count variable.}

\item{popvar}{The person-time (or another exposure) variable.}

\item{treatvar}{The treatment group indicator dummy (0 for controls, 1 for treated units).}

\item{postvar}{The post-period indicator dummy (0 for all time points in the pre-intervention period, 1 for all time points in the post-period)}

\item{timevar}{The time variable (can be coded on any scale).}

\item{idvar}{The panel id variable.}

\item{data}{A long-form panel dataset containing the supplied variables.}

\item{iter}{The number of subsampling iterations. Defaults to 10, but 500-1000 are usually recommended.}
}
\value{
Returns a list containing the following elements:
\code{$Resdata}: a data frame containing the results.
\code{$ECDF}: the empirical cumulative distribution function.
\code{$Effects}: a long-form data frame containing the estimated placebo effects.
\code{$Treat.ratio}: a data frame containing the post/pre-RMSE ratio for the treated unit.
\code{$supp_stats}: a data frame containing supplementary statistics (p-values etc).
}
\description{
\code{iddplacebo} estimates placebo studies on untreated units and produces pseudo p-values based on the empirical distribution of post/pre-RMSE ratios.
}
\examples{
\dontrun{
data(simpanel)
placebo.out <- iddplacebo(eventvar="y",
             popvar="pop",
             idvar="age",
             timevar="time",
             postvar="post",
             treatvar="treat",
             data=simpanel,
             iter=50)
}
}
