% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iddplacebo.spaghetti.R
\name{iddplacebo.spaghetti}
\alias{iddplacebo.spaghetti}
\title{Plot the estimated effects from the placebo studies.}
\usage{
iddplacebo.spaghetti(x, mult = 1e+05, rm = TRUE, alpha = 0.05)
}
\arguments{
\item{x}{Name of the \code{iddplacebo} output object.}

\item{mult}{Multiplier for the rates. Defaults to 100000.}

\item{rm}{Remove placebo studies with poor pre-intervention fit? Default=TRUE.}

\item{alpha}{controls the shaded significance region on the plot. Defaults to 0.05.}
}
\value{
Returns a ggplot containing the estimated effects from the placebo studies along with the results from the main analysis.
}
\description{
\code{iddplacebo.spaghetti} takes an output object after \code{iddplacebo} and plots the estimated placebo effects using \code{ggplot2}.
}
\examples{
\dontrun{
data(simpanel)
placebo.out <- iddplacebo(eventvar="y",
             popvar="pop",
             idvar="age",
             timevar="time",
             postvar="post",
             treatvar="treat",
             data=simpanel,
             iter=50)
plot.out <- iddplacebo.spaghetti(placebo.out)
}
}
