% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ind.tau.R
\name{ind.tau}
\alias{ind.tau}
\title{Solves the basic individual-based serial interval model}
\usage{
ind.tau(Tg = 5, R0 = 2.6, Nmax = 30)
}
\arguments{
\item{Tg}{generation time}

\item{R0}{the basic reproductive number}

\item{Nmax}{the maximum number if infections in a cluster}
}
\value{
Return a list of infectors and times of infection.
The function itself is not yet implemented
}
\description{
Simulates a sequence of infectin events from individuals
with infections drawn from simple offspring and serial
interval distributions.
}
\examples{
ind.tau()
}

