% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_tau.R
\name{ind_tau}
\alias{ind_tau}
\title{Solves the basic individual-based serial interval model}
\usage{
ind_tau(Tg = 5, R0 = 2.6, Nmax = 30)
}
\arguments{
\item{Tg}{generation time}

\item{R0}{the basic reproductive number}

\item{Nmax}{the maximum number if infections in a cluster}
}
\value{
Returns a list of times of infection and indices.
The position in the list of infection times corresponds to
the index given in the second argument.

XXXX This function not fully implemented yet XXXX
}
\description{
Simulates a sequence of individual infection events.
Infections are drawn from simple offspring and serial
interval distributions.
}
\examples{
ind.tau()
}
