% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laiv_cohort.r
\name{laiv_cohort}
\alias{laiv_cohort}
\title{This function is supposed to return simulated cohort studies
of young children in the UKL that we can use to evaluate the
possible impact of repeated LAIV vaccination on the population
The default arguments are currently a little clumsy}
\usage{
laiv_cohort(
  np = 100,
  ny = 6,
  inf_sus = 0.3333,
  an_attack = rep(0.15, ny),
  an_vacc = c(rep(0, 2), rep(0.5, 4))
)
}
\description{
This function is supposed to return simulated cohort studies
of young children in the UKL that we can use to evaluate the
possible impact of repeated LAIV vaccination on the population
The default arguments are currently a little clumsy
}
