% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_dash_data_uk.r
\name{load_uk_cov_data}
\alias{load_uk_cov_data}
\title{Look here to figure out field names}
\usage{
load_uk_cov_data(
  at = "region",
  met1 = "newCasesBySpecimenDate",
  met2 = "",
  met3 = "",
  met4 = ""
)
}
\description{
Look here to figure out field names
}
