% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_update.r
\name{mh_update}
\alias{mh_update}
\title{Proposes parameter updates for an MCMC algorithm}
\usage{
mh_update(pt, fmask = 1:(dim(pt)[1]))
}
\arguments{
\item{pt}{a table of parameter names, values and min and maxes. 
See Details.}

\item{fm}{a vector of strings of the parameters within the table 
to be fitted.
The default value is to assume all the parameters are to be fitted}
}
\value{
A vector of proposed parameter values. The vector is the 
same length as ptab so it copies in the values of the parameters 
that are not being updated.
}
\description{
Uses either linear or log-linear random walks to propose 
multivariate jumps in parameter space.
}
\details{
The table of parameter names must have the correct columns for
this to work...
}
