% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_fig_proof.r
\name{pdf_fig_proof}
\alias{pdf_fig_proof}
\title{Opens a pdf device with sensible options}
\usage{
pdf_fig_proof(findex = 1, file = paste("./pdf_sub_", findex, ".pdf", sep =
  ""), pw = 7, ph = 7, textscale = 0.6, xpd = NA)
}
\arguments{
\item{findex}{is a name index to make it easier to generate many plots
using the same basic file stem.}

\item{file}{the filename for the pdf}

\item{pw}{plotting area width in cm}

\item{ph}{plotting area height in cm}

\item{textscale}{scale factor for axis labels}
}
\description{
Starts a new pdf device with the width and height of the plotting
area specified in real distance units and with a specified margin
measured from the plotting area to the edge of the pdf bounding box.
This should make it easier to compose multi part figure. The device
that gets opened with this still needs to be closed.
}
