% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo.ciplot.R
\name{placebo.ciplot}
\alias{placebo.ciplot}
\title{Plot time-varying placebo-based confidence intervals.}
\usage{
placebo.ciplot(object, mult = 1e+05)
}
\arguments{
\item{object}{An R object obtained using \code{placeboci}.}

\item{mult}{A multiplier for the rates, defaults to 100000.}
}
\value{
Returns a plot containing time-specific placebo-based confidence intervals.
}
\description{
\code{placebo.ciplot} takes the results from \code{placeboci} and plots the time-varying confidence intervals for the effect.
}
\examples{
\dontrun{
data(simpanel)
placebo.out <- iddplacebo(eventvar="y",
             popvar="pop",
             idvar="age",
             timevar="time",
             postvar="post",
             treatvar="treat",
             data=simpanel,
             iter=50)
ci.out <- placeboci(placebo.out, alpha=0.05)
placebo.ciplot(ci.out)
}
}
