% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo.pplot.R
\name{placebo.pplot}
\alias{placebo.pplot}
\title{Plot time-varying placebo-based p-values.}
\usage{
placebo.pplot(object, alpha = 0.05)
}
\arguments{
\item{object}{An R object obtained using \code{placeboci}.}

\item{alpha}{Alpha value for the significance region on the plot, defaults to 0.05.}
}
\value{
Returns a plot containing time-specific placebo-based confidence intervals.
}
\description{
\code{placebo.pplot} takes the results from \code{placeboci} and plots the time-varying p-values.
}
\examples{
\dontrun{
data(simpanel)
placebo.out <- iddplacebo(eventvar="y",
             popvar="pop",
             idvar="age",
             timevar="time",
             postvar="post",
             treatvar="treat",
             data=simpanel,
             iter=50)
ci.out <- placeboci(placebo.out, alpha=0.05)
placebo.pplot(ci.out)
}
}
