% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeboci.R
\name{placeboci}
\alias{placeboci}
\title{Estimate placebo-based confidence intervals.}
\usage{
placeboci(object, alpha = 0.05, mult = 100000)
}
\arguments{
\item{object}{An R object containing the results from a set of placebo studies obtained using \code{iddplacebo}.}

\item{alpha}{The significance level, defaults to 0.05 for 95 percent CI.}

\item{mult}{Multiplier for the rate, defaults to 100000.}
}
\value{
The code prints the CI for the average post-intervention effect, and returns a data frame containing the time-varying CI. To be used with \code{iddplacebo.ciplot} or \code{iddplacebo.pplot}.
}
\description{
\code{placeboci} takes the results from a set of placebo studies obtained using \code{iddplacebo} and estimates non-parametric placebo-based confidence intervals for a selected alpha level.
}
\examples{
\dontrun{
data(simpanel)
placebo.out <- iddplacebo(eventvar="y",
             popvar="pop",
             idvar="age",
             timevar="time",
             postvar="post",
             treatvar="treat",
             data=simpanel,
             iter=50)
ci.out <- placeboci(placebo.out, alpha=0.05)
}
}
