% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_practical.R
\name{plot_incidence_all}
\alias{plot_incidence_all}
\title{plots the incidence for a given country (all years)}
\usage{
plot_incidence_all(flu_data, country_code = "ISR")
}
\arguments{
\item{flu_data}{matrix with flu incidence data}

\item{country_code}{character vector of length 1: 3-letter country code}
}
\value{
ggplot object
}
\description{
\code{plot_incidence_all} plots the incidence for a given country (all years)
}
