% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_practical.R
\name{plot_likelihood_profile}
\alias{plot_likelihood_profile}
\title{plot a likelihood profile for R_0 given data}
\usage{
plot_likelihood_profile(likelihood_profile_output)
}
\arguments{
\item{likelihood_profile_output}{data frame returned by 
\code{likelihood_profile_seir}}
}
\value{
ggplot object
}
\description{
\code{plot_likelihood_profile} plots the log likelihood of the model 
given the data, for a range of R_0. The other parameters are fixed
}
