% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simpanel}
\alias{simpanel}
\title{A toy panel dataset of five age groups, with two poorly fitting and two good fitting controls.}
\format{A small toy dataset containing the following variables:
\describe{
  \item{age}{age group, used as id variable}
  \item{treat}{treatment unit dummy variable}
  \item{y}{outcome variable, n events}
  \item{time}{time variable}
  \item{post}{post-intervention dummy}
  \item{pop}{population denominator}
  \item{inc}{incidence rate (y/pop), not used}
}}
\usage{
simpanel
}
\description{
A toy panel dataset of five age groups, with two poorly fitting and two good fitting controls.
}
\keyword{datasets}
