% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_practical.R
\name{simple_comp_seir}
\alias{simple_comp_seir}
\title{simplified version of SEIR model with default inputs}
\usage{
simple_comp_seir(R_0)
}
\arguments{
\item{R_0}{numeric vector of length 1: basic reproduction number}
}
\value{
A list of two elements. The first is the incidence of infection
and the second is the timepoints to which the incidence refers.
}
\description{
\code{simple_comp_seir} is a simplified version of \code{comp_seir}
with default inputs
}
