% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allData.R
\docType{data}
\name{smeDf}
\alias{smeDf}
\title{Simulation model example data frame}
\description{
This file contains synthetic data on the location of farms in Kent, UK.
These are not real locatios but they have a similar density distribution.
These data are intended to be used in a teaching example about
simulation models of infectious disease.
}
\author{
My Name \email{blahblah@roxygen.org}
}
\references{
\url{data_blah.com}
}

