% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_practical.R
\name{solve_seir_model}
\alias{solve_seir_model}
\title{solve SEIR model for incidence}
\usage{
solve_seir_model(R_0, latent_period, infectious_period, N, I_0, n_weeks)
}
\arguments{
\item{R_0}{numeric vector of length 1: basic reproduction number}

\item{latent_period}{numeric vector of length 1: latent period}

\item{infectious_period}{numeric vector of length 1: infectious period}

\item{N}{numeric vector of length 1: population size}

\item{I_0}{numeric vector of length 1: initial number of infectious individuals}

\item{n_weeks}{numeric vector of length 1: number of weeks for which to solve the model}
}
\value{
data frame with predicted points and original data
}
\description{
\code{solve_seir_model} solves the SEIR model for the weekly incidence
}
