% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_practical.R
\name{solve_seir_wrapper}
\alias{solve_seir_wrapper}
\title{solve SEIR model for incidence, specifying default parameters except for 
\code{R_0} and \code{n_weeks}}
\usage{
solve_seir_wrapper(R_0, n_weeks)
}
\arguments{
\item{R_0}{numeric vector of length 1: basic reproduction number}

\item{n_weeks}{numeric vector of length 1: number of weeks for which to solve the model}
}
\value{
data frame with predicted points and original data
}
\description{
\code{solve_seir_wrapper} solves the SEIR model for the weekly incidence,
specifying default parameters except for \code{R_0} and \code{n_weeks}
}
