% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allFuncs.R
\name{srg.param.propose.update}
\alias{srg.param.propose.update}
\title{Proposes parameter updates for an MCMC algorithm}
\usage{
srg.param.propose.update(ptab, fmask = 1:(dim(ptab)[1]))
}
\arguments{
\item{ptab}{a table of parameter names, values and min and maxes. See Details.}

\item{fmask}{a vector of strings of the parameters within the table to be fitted.
The default value is to assume all the parameters are to be fitted}
}
\value{
A vector of proposed parameter values. The vector is the same length as ptab
so it copies in the values of the parameters that are not being updated.
}
\description{
Uses either linear or log-linear random walks to propose multivariate
jumps in parameter space.
}

