% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_practical.R
\name{subset_incidence_data}
\alias{subset_incidence_data}
\title{subset incidence data for weeks of interest}
\usage{
subset_incidence_data(incidence_data, current_week, n_weeks_prior)
}
\arguments{
\item{incidence_data}{data frame extracted by \code{extract_incidence}}

\item{current_week}{numeric vector of length 1: week number of the current week}

\item{n_weeks_prior}{numeric vector of length 1: number of weeks' previous data to use}
}
\value{
data frame for weeks of interest
}
\description{
\code{subset_incidence_data} takes the incidence data and subsets it for the weeks of interest.
}
