% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zika_invasion.R
\name{zika_invasion}
\alias{zika_invasion}
\title{Illustration of possible zika invasion dynamics}
\usage{
zika_invasion(
  its = 10,
  file = "screen",
  N = 1e+06,
  seed = 19283719,
  highlights = c(1, 2, 3),
  highcolors = c("red", "magenta", "cyan"),
  maxts = 360 * 2,
  R0axes = 0:4,
  incaxes = seq(0, 3.5, 0.5)
)
}
\arguments{
\item{its}{number of iterations for the second population}

\item{file}{name of file to write pdf or "screen"}

\item{N}{size of population}
}
\description{
Simulates a stochastic compartmental model to reproduce an
illustration of how two very similar communities might have
very different initial epidemics.
}
\examples{
zika_invasion(its=100,file="~/tmp/zika_example.pdf")
zika_invasion(its=100,file="screen",N=1000000)
}
