% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_react_funcs.r
\name{add.CIs}
\alias{add.CIs}
\title{Define a function to take a table and add in binomial estimates and
CIs}
\usage{
add.CIs(tab, method = "wilson")
}
\arguments{
\item{tab}{The input table for which CIs are needed.}

\item{method}{The optimisation mthod needed, as passed to prop CI}
}
\description{
Define a function to take a table and add in binomial estimates and
CIs
}
