% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allFuncs.R
\name{comp.seir}
\alias{comp.seir}
\title{Solves the SEIR model}
\usage{
comp.seir(De = 1.48, Tg = 2.6, R0 = 1.8, N = 6800000, I0 = 10,
  dt = 1, R1 = 1, t1 = 999, R2 = 1, t2 = 9999, noTimeSteps = 10,
  noReals = 1, deterministic = TRUE)
}
\arguments{
\item{De}{duration latent}

\item{Tg}{generation time}

\item{R0}{Basic reproductive number}

\item{N}{size of the population}

\item{I0}{number of initial infectious individuals}

\item{dt}{time step}

\item{R1}{second value for the reproductive nuber}

\item{t1}{time of change to the second R value}

\item{R2}{third value for the reproductive number}

\item{t2}{time of change to the third reproductive number}

\item{noTimeSteps}{number of timesteps}

\item{noReals}{number of realizations, NOT YET IMPLEMENTED}

\item{deterministic}{mode of solution, NOT YET IMPLEMENTED}
}
\value{
A list of two elements. The first is the incidence of infection and the
second is the timepoints to which the incidence refers.
}
\description{
Creates both deterministic and stochastic solutions to a common
epidemic model.
}
\examples{
comp.seir()
comp.seir(R0=0.999)
}

