% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fluIliCountryData}
\alias{fluIliCountryData}
\title{Forecast model example data frame}
\description{
These are influenza-like-illness data gathered from many countries across the 
world and curated by the WHO as FluID.
}
\references{
These incidence data are from publically available WHO syndromic data
\url{https://www.who.org}
}
\author{
Steven Riley \email{sr@stevenriley.net}
}
