\name{idd-package}
\alias{idd-package}
\alias{idd}
\docType{package}
\title{
\packageTitle{idd}
}
\description{
\packageDescription{idd}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{idd}
\packageIndices{idd}
This package provides a variety of functions to help investigate infectious disease dynamics. The functions fall into }
\author{
\packageAuthor{idd}

Maintainer: \packageMaintainer{idd}
}
\references{
~~ Literature or other references for background information ~~
}
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%% ~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
