% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_fig_proof.R
\name{pdf_fig_proof}
\alias{pdf_fig_proof}
\title{A function to setup a clean pdf file for a single subplot}
\usage{
pdf_fig_proof(
  findex = 1,
  file = paste("./pdf_sub_", findex, ".pdf", sep = ""),
  pw = 7,
  ph = 7,
  textscale = 0.6,
  xpd = NA
)
}
\arguments{
\item{plotheight}{Is a parameter}

\item{margin}{Is a parameter}

\item{pdfheight}{Is a parameter}

\item{plotwidth}{Is a parameter}

\item{posplot}{Is a parameter}
}
\description{
A function to setup a clean pdf file for a single subplot
}
\examples{
pdf_fig_proof()
}
