% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_practical.R
\name{plot_incidence}
\alias{plot_incidence}
\title{plots the incidence}
\usage{
plot_incidence(incidence_data, log_scale = FALSE)
}
\arguments{
\item{incidence_data}{data frame extracted by \code{extract_incidence}}

\item{log_scale}{logical vector of length 1: if TRUE, plot on log scale, 
otherwise plot on linear scale}
}
\value{
ggplot object
}
\description{
\code{plot_incidence} plots the incidence extracted by \code{extract_incidence}
}
