% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smeDf}
\alias{smeDf}
\title{Simulation model example data frame}
\description{
This object contains synthetic data on the location of farms in Kent, UK.
These are not real locations but they have a similar density distribution.
These data are intended to be used in a teaching example about
simulation models of infectious disease.
}
\references{
These synthetic data have been constructed to have a similar
distribution of spatial density as the actual distribution in this part of
the UK.
\url{https://www.gov.uk/government/statistical-data-sets/structure-of-the-livestock-industry-in-england-at-december}
}
\author{
Steven Riley \email{sr@stevenriley.net}
}
